+++ Screen_Regions.py
@@ -13,6 +13,47 @@ Author: sumzer0@yahoo.com
 """
 
 
+def reg_scale_for_station(region, w: int, h: int) -> [int, int, int, int]:
+    """ Scale a station services region based on the target resolution.
+    This is performed because the tables on the station services screen do
+    not increase proportionally with the screen size. The width changes with
+    the screen size, the height does not change based on the screen size
+    height, but on the screen width and the position stays consistent to the
+    center of the screen.
+    To calculate the new region height, we take the initial region defined at
+    1920x1080 and scale up the height based on the target width and apply the
+    new proportion against the center line.
+    @param h: The screen height in pixels
+    @param w: The screen width in pixels
+    @param region: The region at 1920x1080
+    @return: The new region in %
+    """
+    ref_w = 1920
+    ref_h = 1080
+
+    # Calc the x and y scaling.
+    x_scale = w / ref_w
+    y_scale = h / ref_h
+
+    # Determine centre of the region
+    reg_avg = 0.5
+    # This alternate method below is based on the centre of the region instead of the centre of screen.
+    # This will generally NOT work for station screens that are obviously centred vertically.
+    # reg_avg = (region['rect'][1] + region['rect'][3]) / 2
+
+    # Recalc the region as a % above and below the center line.
+    pct_abv = (reg_avg - region['rect'][1]) * x_scale / y_scale
+    pct_blw = (region['rect'][3] - reg_avg) * x_scale / y_scale
+
+    # Apply new % to the center line.
+    new_rect1 = reg_avg - pct_abv
+    new_rect3 = reg_avg + pct_blw
+
+    # Return the update top and bottom Y percentages with the original X percentages.
+    new_reg = {'rect': [region['rect'][0], new_rect1, region['rect'][2], new_rect3]}
+    return new_reg
+
+
 def size_scale_for_station(width: int, height: int, w: int, h: int) -> (int, int):
     """ Scale an item in the station services region based on the target resolution.
     This is performed because the tables on the station services screen do
@@ -66,7 +107,7 @@ class Screen_Regions:
 
         self.reg = {}
         # regions with associated filter and color ranges
-        # The rect is top left x, y, and bottom right x, y in fraction of screen resolution
+        # The rect is [L, T, R, B] top left x, y, and bottom right x, y in fraction of screen resolution
         self.reg['compass']   = {'rect': [0.33, 0.65, 0.46, 1.0], 'width': 1, 'height': 1, 'filterCB': self.equalize,                                'filter': None}
         self.reg['target']    = {'rect': [0.33, 0.27, 0.66, 0.70], 'width': 1, 'height': 1, 'filterCB': self.filter_by_color, 'filter': self.orange_2_color_range}   # also called destination
         self.reg['target_occluded']    = {'rect': [0.33, 0.27, 0.66, 0.70], 'width': 1, 'height': 1, 'filterCB': self.filter_by_color, 'filter': self.target_occluded_range}