+++ ChangeLog.md
@@ -1,3 +1,20 @@
+# 7/3/2025 Stumpii: Small improvements.
+ - Added Russian localisation file.
+ - Added language selection for OCR.
+ - Added localisation to nav panel.
+ - Added TCE Integration class.
+ - Added OCR to station services screen check.
+ - Added fix for Robigo issue not selecting system in gal map.
+ - Improved debug for gal map.
+ - Separated Station Services and Construction Services because OCR only works for Station Services (Construction Services are not as slow to open).
+ - Moved docking cleanup to Nav Panel class.
+ - Added OCR test class.
+ - Test code for gal map.
+ - Prevent gal map being opened twice for no reason.
+ - Added region scaling for stations.
+ - Added EDNavigationPanel.py for navigation panel code.
+ - Disabled build-executable.yml as it always fails.
+
 # 6/1/2025 Stumpii: Bug fix.
  - Removed 'deprecated' warning that does not work in old python versions.